multiple fixes + repaired transfer learning experiments

+++ experiments/hmmvsrnn_reco/b_preprocess.py
@@ -126,12 +126,12 @@ def skel_feat(pose_seq):
 def bgr_feats(frame_seq, pose2d_seq):
     d = len(pose2d_seq)
     cropw, croph = crop_size
-    crop_seq = np.empty((d, 2, cropw + 30, croph + 30, 3), dtype=frame_seq[0].dtype)
+    crop_seq = np.empty((d, 2, cropw + 30, croph + 30, 3), dtype=frame_seq.dtype)
     for c, f, p in zip(crop_seq, frame_seq, pose2d_seq):
         c[1] = crop(f, p[joints.WristLeft], (cropw + 30, croph + 30))
         c[0] = crop(f, p[joints.WristRight], (cropw + 30, croph + 30))
 
-    output = np.empty((d, 2, cropw, croph), dtype=frame_seq[0].dtype)
+    output = np.empty((d, 2, cropw, croph), dtype=np.float32)
     output[:, 0] = fix_contrast(crop_seq[:, 0])[:, 15:-15, -15:15:-1]
     output[:, 1] = fix_contrast(crop_seq[:, 1])[:, 15:-15, 15:-15]
